BD/streamlit/requirements.txt
streamlit
plotly
pandas
kafka-python
joblib
scikit-learn